+++ sc-carbon-footprint/src/GovernancePoCRNet.sol
@@ -6,7 +6,7 @@ import "./Governance.sol";
 
 contract GovernancePoCRNet is Governance {
 
-    function getConstantValue(uint key)
+    function getConstantValue(uint256 key)
         internal
         pure
         override
